feat: mejora la validación de datos en los DTOs

+++ src/main/java/com/oo2/grupo17/dtos/ContactoDto.java
@@ -20,7 +20,7 @@ public class ContactoDto {
     @NotNull(message = "El móvil es obligatorio")
     @Min(value = 1000000000, message = "El móvil debe tener al menos 10 dígitos")
     @Max(value = 9999999999L, message = "El móvil no debe superar los 10 dígitos")
-    private int movil;
+    private long movil;
 
     @Min(value = 10000000, message = "El teléfono debe tener al menos 8 dígitos")
     @Max(value = 99999999, message = "El teléfono no debe superar los 8 dígitos")
